Refactor globe covering tiles into a separate file
