+++ src/geo/transform_interface.ts
@@ -9,6 +9,7 @@ import {Terrain} from '../render/terrain';
 import {PointProjection} from '../symbol/projection';
 import {MapProjectionEvent} from '../ui/events';
 import type {ProjectionData} from './projection/projection_data';
+import {IntersectionResult} from '../util/primitives';
 
 export type CoveringZoomOptions = {
     /**
@@ -301,9 +302,9 @@ export interface IReadonlyTransform extends ITransformGetters {
     getVisibleUnwrappedCoordinates(tileID: CanonicalTileID): Array<UnwrappedTileID>;
 
     /**
-     * Return all tile coordinates that could cover this transform for a covering
-     * zoom level, ordered by ascending distance from camera.
-     * @param options - the options
+     * Returns a list of tile coordinates that when rendered cover the entire screen at an optimal detail level.
+     * Tiles are ordered by ascending distance from camera.
+     * @param options - Additional options - min & max zoom, terrain presence, etc.
      * @returns Array of OverscaledTileID. All OverscaledTileID instances are newly created.
      */
     coveringTiles(options: CoveringTilesOptions): Array<OverscaledTileID>;
@@ -496,3 +497,15 @@ export interface IReadonlyTransform extends ITransformGetters {
  * A transform is cloneable, which is used when a given map state must be retained for multiple frames, mostly during symbol placement.
  */
 export interface ITransform extends IReadonlyTransform, ITransformMutators {}
+
+export interface ITileVisibilityProvider {
+    /**
+     * A function that returns whether the tile is visible under the current transform.
+     * The computation needs not be completely accurate, but no tile that is even partially visible can be returned as not visible.
+     * @param x - Tile X.
+     * @param y - tile Y.
+     * @param z - Tile zoom.
+     * @returns 0 is not visible, 1 if partially visible, 2 if fully visible.
+     */
+    isTileVisible(x: number, y: number, z: number): IntersectionResult;
+}
